requirements.txt
Django==5.2
djangorestframework==3.16.0
firebase-admin==6.9.0
django-jazzmin==3.0.1
gunicorn==23.0.0